requirements.txt
google-generativeai
haystack-ai
sentence-transformers
beautifulsoup4
python-dotenv
requests